requirements.txt
tk
customtkinter
Pillow
matplotlib
requests
python_dotenv
entsoe-py